build-m68knommu-linux: introduce flavor for CPU/system variants
For any one CPU architecture we may actually have different types of
system configurations we want to build for and test. Introduce the
concept of a "flavor" of a CPU type. For example we may want to build
an M68K based system with or without an MMU. Or a system using ELF
format executables or "flt" format executables.

So flavors would look like this:

    m68knommu-flt
    m68knommu-elf
    m68k-flt
    m68k-elf
    armnommu-flt
    armnommu-elf
    ...

So the presence of an MMU is either "" or "nommu". The CPU and MM
configuration is commonly concatenated within linux, so stick to that
style with no '-' separator.

Executable type "flt" or "elf".

There may end up being many more variations.

Signed-off-by: Greg Ungerer <gerg@linux-m68k.org>