configs: create sub-directory for configuration files
Move the package configuration files into their own sub-directory.

With the introduction of flavors there will end up being a number of
specific configuration files for some packages like uClibc and busybox,
and maybe even linux itself in the future. So it makes sense to group
them all together out of the way of the main build scripts.

Signed-off-by: Greg Ungerer <gerg@linux-m68k.org>