build*: limit number of parallel jobs
Building with gcc-12 and linux-6.1 now seems to let the make parallel
build server run crazy and just spawn too many jobs. Even on a large
workstation it will quickly start swapping and flailing badly while
building the kernel.

Limit the number of jobs to an arbitary 32. This could almost ceratinly
be smarter based on how many physical CPU cores there is present. But
that is a fix for another time.

Signed-off-by: Greg Ungerer <gerg@kernel.org>