build-armnommu-linux-uclibc-elf: build for ELF format binaries on armnommu
Add a new build script for ELF format binaries running on ARM nommu
systems. This follows the same uldso based loader as used on other ELF
based nommu target builds. Great thing adout ARM is that it already supports
the ELF-FDPIC loader, so no new kernel changes required for that. It is
a little buggy though, so some patches required to make it work.

Signed-off-by: Greg Ungerer <gerg@kernel.org>