build-armnommu-flt: change elf2flt patch set
Change the set of patches used to fix elf2flt to use a different method
of fixing the ARM.exidx problem. The problem with the mainline elf2flt
source is that it is broken on a number of other architectures and that
stems back to the original ARM.exidx fix. The new patches revert the
offending changes and then apply a different solution.

This change also takes the opportunity to use the elf2flt proposed
mainline fix for the binutils-2.34 breakage.

All up this should result in no actual difference in operational
behavior of the built system. But it will make it easier to use the
mainline elf2flt sources on other architectures (like m68k).

Signed-off-by: Greg Ungerer <gerg@kernel.org>