build: force controlling tty on console shell
Update the inittab entry for the console shell so that it will set the
controlling tty of that shell process. This relies on the busybox option
that uses a leading '-' to signify setting the controlling tty on an
inittab spawned process.A This means that CTRL-C now works on the
console command shell.

Signed-off-by: Greg Ungerer <gerg@kernel.org>