build: nommu: elf: update to using uldso v1.0.4
Update the build scripts for the nommu targets that use an ELF user
space to use the latest version of uldso, v1.0.4. This newest version
does not rely on the ELF-FDPIC register setups at process start time.
Instead they only rely on the traditonal ELF stack setup to carry out
the dynamic relocation.

Signed-off-by: Greg Ungerer <gerg@kernel.org>