build: m68knommu-elf: use uclibc targeted toolchain
Select the target toolchain as "m68k-linux-uclibc" to be consistent
with the other architecture builds. Technically this is more accurate
any way - this is building a uClibc target toolchain.

The main implication is that the run-time dynamic loaded is expected
to be named "ld-uClibc.so.0", and not the glibc "ld.so.1".

Signed-off-by: Greg Ungerer <gerg@kernel.org>