build-nommu-elf: switch to using uClibc's ldso
The run-time dynamic loading of ELF format binaries on noMMU systems is
not really a special case of dynamic linking, and the uClibc ldso is
quite capable of doing the job.

Add an "ELF" binary file type configuration option to uClibc that
chooses use of binfmt_elf_fdpic loader ELF binaries. This replaces the
some what odd and wrong choice of having to pick flat format binaries in
this scenario. The ELF format binary choice will select the approiate
other uClibc options (like DOPIC and STATIC_PIE) and also build the ldso
dynamic link loader. This replaces the use of uldso which is no logner
required.

Signed-off-by: Greg Ungerer <gerg@kernel.org>