build: riscnommu: fix application dumps
uClibc-ng-1.0.48 enabled CTOR/DTOR support for RiSC-V nommu mode. This
causes applications to crash/dump on exit. So they run normally to
completion, but then get a signal 11 and exit. The cause appears to be
that the __do_global_dtors_aux() code path in the crt start up code
is not PIC, and so does not have appropriate relocations. For now work
around this problem by reverting that uClibc-ng change - until a proper
fix has been determined.

This also exposed a bug in the linux kernel's flat loader that is not
fixed in mainline kernels yet. The flat loader is processing possible
shared library entries that do not exist for RISC-V nommu. Apply the fix
here with appropriate patch.

Signed-off-by: Greg Ungerer <gerg@kernel.org>