build: riscnommu: drop uClibc CTOR/DTOR patch
uClibc-ng-1.0.49 modified the UCLIBC_CTOR_DTOR config option so that
it is now boolean selectable. By disabling this option we no longer get
applications core dumping on exit. We no longer need to patch this
into uClibc to be able to disable it.

Remove the local CTOR/DTOR patch and modify the uClibc configurations
for RISC-V nommu targets to disable UCLIBC_CTOR_DTOR.

Signed-off-by: Greg Ungerer <gerg@kernel.org>