build: drop build and use of opensbi for RISC-V 64bit
Modern versions of qemu seem to include a usable version of the opensbi
loader for generic virt RISC-V 64bit platforms. So no need to build and
run a local one. Remove build of it and remove its command line use
when running qemu (really a documentation change for this part).

Signed-off-by: Greg Ungerer <gerg@kernel.org>