build: use current binutils for cortex build
The cortex platform build is broken when using binutils-2.44 and newer.
The breakage is observed when compiling busybox, producing the
following error:

    toolchain/arm-uclinuxeabi/bin/ld.real: toolchain/lib/gcc/arm-uclinuxeabi/15.2.0/thumb/libgcc.a(_aeabi_uldivmod.o)(__aeabi_uldivmod): Unknown destination type (ARM/Thumb) in networking/lib.a(ping.o)
    networking/lib.a(ping.o): in function `print_stats_and_exit':
    ping.c:(.text.print_stats_and_exit+0x48): dangerous relocation: unsupported relocation

Bisecting through the binutils-gdb git tree the problem was found to be
due to the changes in commit 31ed3a9d6914 ("arm: Do not insert stubs
needing Arm code on Thumb-only cores."). Reverting that change fixes the
busybox build problem.

Switch the cortex build back to using the modern binutils (currently
2.45.1) and reverting that change.

Signed-off-by: Greg Ungerer <gerg@kernel.org>